chore: refactor the province test codes
