+++ app/__tests__/provinces/index.js
@@ -1,0 +1,7 @@
+const createInstances = require('./createInstances')
+const updateInstances = require('./updateInstances')
+
+module.exports = {
+  createInstances,
+  updateInstances
+}
