+++ app/jest.config.js
@@ -4,6 +4,9 @@ module.exports = {
   moduleFileExtensions: ['js'],
   testPathIgnorePatterns: [
     '/node_modules/',
-    '__tests__/classInitialization/checkClass.js'
+    '__tests__/classInitialization/checkClass.js',
+    '__tests__/provinces/createInstances.js',
+    '__tests__/provinces/updateInstances.js',
+    '__tests__/provinces/index.js'
   ]
 }
